improve api

+++ src/main/kotlin/com/wutsi/platform/rtm/model/ChatUser.kt
@@ -2,11 +2,11 @@ package com.wutsi.platform.rtm.model
 
 data class ChatUser(
     val id: String = "",
-    val createdAt: Int = -1,
+    val createdAt: Long = -1,
     val firstName: String? = null,
     val lastName: String? = null,
     val imageUrl: String? = null,
-    val lastSeen: Int? = null,
-    val updatedAt: Int? = null,
+    val lastSeen: Long? = null,
+    val updatedAt: Long? = null,
     val metadata: Map<String, Any>? = null
 )
